requirements.txt
async-generator==1.10
attrs==21.4.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
distlib==0.3.4
docutils==0.18.1
filelock==3.4.0
Flask==2.0.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
imagesize==1.3.0
itsdangerous==2.0.1
jdcal==1.4.1
Jinja2==3.0.2
lxml==4.6.4
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.26.1
mysql-connector-python==8.0.27
mysqlclient==2.1.0
openpyxl==2.1.4
outcome==1.1.0
packaging==21.3
Pillow==8.4.0
platformdirs==2.4.0
protobuf==3.19.1
pycparser==2.21
Pygments==2.10.0
pyodbc==4.0.32
pyOpenSSL==21.0.0
pyparsing==3.0.6
PyPDF2==1.26.0
pyperclip==1.8.2
python-docx==0.8.11
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.28
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
virtualenv==20.10.0
Werkzeug==2.0.2
wsproto==1.0.0
WTForms==3.0.0
